feat(disclose): deplace le setter ungroup sur le disclose
