+++ src/core/script/collapse/collapse-selector.js
@@ -2,6 +2,5 @@ import ns from '../api/utilities/namespace.js';
 
 export const CollapseSelector = {
   COLLAPSE: ns.selector('collapse'),
-  COLLAPSING: ns.selector('collapsing'),
-  GROUP: ns.attr('group')
+  COLLAPSING: ns.selector('collapsing')
 };
