+++ src/core/script/disclosure/disclosure-selector.js
@@ -1,5 +1,6 @@
 import ns from '../api/utilities/namespace.js';
 
 export const DisclosureSelector = {
-  PREVENT_CONCEAL: ns.attr.selector('prevent-conceal')
+  PREVENT_CONCEAL: ns.attr.selector('prevent-conceal'),
+  GROUP: ns.attr('group')
 };
