+++ src/core/script/disclosure/disclosures-group.js
@@ -1,6 +1,7 @@
 import { Instance } from '../api/modules/register/instance.js';
 import { DisclosureEmission } from './disclosure-emission.js';
 import { completeAssign } from '../api/utilities/property/complete-assign.js';
+import { DisclosureSelector } from './disclosure-selector.js';
 
 class DisclosuresGroup extends Instance {
   constructor (disclosureInstanceClassName, jsAttribute) {
@@ -22,6 +23,7 @@ class DisclosuresGroup extends Instance {
     this.addAscent(DisclosureEmission.RETRIEVE, this.retrieve.bind(this));
     this.addAscent(DisclosureEmission.REMOVED, this.update.bind(this));
     this.descend(DisclosureEmission.GROUP);
+    this._isGrouped = this.getAttribute(DisclosureSelector.GROUP) !== 'false';
     this.update();
   }
 
@@ -47,6 +49,9 @@ class DisclosuresGroup extends Instance {
       },
       get hasFocus () {
         return scope.hasFocus;
+      },
+      set isGrouped (value) {
+        scope.isGrouped = value;
       },
       get isGrouped () {
         return scope.isGrouped;
@@ -159,12 +164,26 @@ class DisclosuresGroup extends Instance {
     return false;
   }
 
+  set isGrouped (value) {
+    const isGrouped = !!value;
+    if (this._isGrouped === isGrouped) return;
+    this._isGrouped = isGrouped;
+    this.setAttribute(DisclosureSelector.GROUP, !!value);
+    this.update();
+  }
+
   get isGrouped () {
     return this._isGrouped;
   }
 
   get canUngroup () {
     return false;
+  }
+
+  mutate (attributesNames) {
+    if (attributesNames.includes(DisclosureSelector.GROUP)) {
+      this.isGrouped = this.getAttribute(DisclosureSelector.GROUP) !== 'false';
+    }
   }
 
   apply () {}
